# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SoftClans API",
        "description": "SoftClans API Portal is a platform for documentation of all our APIs",
        "contact": {
            "email": "ronohhilla@gmail.com"
        },
        "license": {
            "name": "SoftClans License 1.0",
            "url": "http://www.softclans.co.ke"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:91/myself/api_portal/api/v1",
            "description": "KIP HOST SERVER"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Read users",
                "description": " ",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Number of records fetched"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "description": " ",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Records fetched successfully"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/user/{user_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Read specified user",
                "description": " ",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Unique identifier of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Records fetched successfully"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "description": " ",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Unique identifier of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Records fetched successfully"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "description": " ",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Unique identifier of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Records fetched successfully"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    }
}